requirements.txt
anyio==4.7.0
certifi==2024.12.14
dnspython==2.7.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
motor==3.7.0
pymongo==4.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.10
six==1.17.0
sniffio==1.3.1
typing_extensions==4.12.2
